+++ examples/AutoBenchmark/generate_readme.py
@@ -39,7 +39,7 @@ which will add to the timing values shown below in actual usage.
 The [digitalWriteFast library](https://github.com/NicksonYap/digitalWriteFast)
 might be an alternative if speed is critical.
 
-**Version**: AceButton v1.9
+**Version**: AceButton v1.9.1
 
 **DO NOT EDIT**: This file was auto-generated using `make README.md`.
 
@@ -113,7 +113,7 @@ number of `TimingStats::update()` calls that were made.
       those calls almost twice as fast as `Encoded4To2ButtonConfig` (38
       microseconds versus 72 microseconds).
 
-**v1.9+**
+**v1.9.1**
 * Upgrade STM32duino Core from 1.9.0 to 2.0.0.
 * Upgrade SparkFun SAMD Core from 1.8.1 to 1.8.3.
 * No significant change in CPU times.